fixed humidity conversion... again

+++ Lab2/lab2_2/main/lab2_2.c
@@ -79,7 +79,7 @@ static esp_err_t read_temperature(float *temperature)
         uint16_t humid_raw = (sensor_data[3] << 8) | sensor_data[4];
         *temperature = -45 + (175 * ((float)temp_raw / 65535));
         // this value is not returned by the function
-        float humidity = 100 * (humid_raw/65535);
+        float humidity = (100 * ((float)humid_raw/65535));
         ESP_LOGI(TAG, "Read temperature: %.2f C", *temperature);
         ESP_LOGI(TAG, "Read Humidity: %.2f %%", humidity);
     } else {
